SDK Nightly build 2026-07-30
List of changes from previous build

## mcu_plus_sdk
Previous: a95a005d98b3 -> Current: c757ab952746

c757ab95274 am26x: ospi: add nand flash phy tuning support
3841734ce7b am261x: usb: add USB benchmark examples

## security_common
Previous: 3f5443787ed9 -> Current: e1d2b0287cfd

e1d2b02 f29h85x: SC: misra

## enet_lld
Previous: ebf42b81f229 -> Current: cfec580f41d0

cfec580f4 am64x,am243x: enet: fix lldp
50572449a am243x,am64x: enet: make lldp example internal
8149fbfd0 am62dx: docs: remove lldp example
8210b61a1 allsoc: docs: remove check in header file

+++ source/security/security_common/drivers/secure_ipc_notify/sipc_notify_src.c
@@ -65,7 +65,7 @@ static inline void SIPC_setMboxConfig(SIPC_Params *params, uint32_t selfCoreId)
     if (selfCoreId == CORE_ID_HSM0_0 )
     {
         /* Define HSM->R5 tx and R5 -> HSM rx queues */
-        for( secMaster = 0 ; secMaster < (uint8_t)(MAX_SEC_CORES_WITH_HSM - 1U); secMaster ++)
+        for( secMaster = 0U ; secMaster < (MAX_SEC_CORES_WITH_HSM - 1U); secMaster ++)
         {
             /*****************************************************************************/
             /* HSM TX queue setup                                                        */
@@ -412,7 +412,7 @@ int32_t SIPC_init(SIPC_Params *params)
     /* check if current core who is doing sipc init is a secure host or not if not.
      * if not then return init failure */
 
-    for( secMaster = 0 ; secMaster < (uint8_t)(MAX_SEC_CORES_WITH_HSM - 1U) ; secMaster ++)
+    for( secMaster = 0U ; secMaster < (MAX_SEC_CORES_WITH_HSM - 1U) ; secMaster ++)
     {
         if((selfCoreId == params->secHostCoreId[secMaster]) || (selfCoreId == CORE_ID_HSM0_0))
         {
@@ -430,7 +430,7 @@ int32_t SIPC_init(SIPC_Params *params)
         }
         else
         {
-            assertFlag = 1;
+            assertFlag = 1U;
         }
     }
     /* Indicates that this core is not a secHost*/
@@ -468,7 +468,7 @@ int32_t SIPC_init(SIPC_Params *params)
                 if(params->coreIdList[core] < MAX_SEC_CORES_WITH_HSM)
                 {
                     /* Mark core as enabled for IPC */
-                    gSIPC_ctrl.isCoreEnabled[params->coreIdList[core]] = 1;
+                    gSIPC_ctrl.isCoreEnabled[params->coreIdList[core]] = 1U;
                 }
                 else
                 {
